refactor: add conditional to FK

+++ src/data/services/Product/DeleteProduct.service.js
@@ -1,8 +1,12 @@
-import { ProductRepository } from '../../database/repositories';
+import {
+  ProductRepository,
+  DocumentRepository,
+} from '../../database/repositories';
 
 export class DeleteProductService {
   async execute({ id_product }) {
     const repository = new ProductRepository();
+    const documentRepository = new DocumentRepository();
 
     const verifyProductExists = await repository.findProductById({
       id_product,
@@ -12,6 +16,17 @@ export class DeleteProductService {
       return {
         error: `Não há nenhum produto registrado com este ID -> ${id_product}.`,
       };
+
+    const verifyFk = await documentRepository.verifyRelationProduct({
+      id_product,
+    });
+
+    if (verifyFk.length > 0) {
+      return {
+        error:
+          'Não foi possível excluir o Produto pois existem Documentos associados.',
+      };
+    }
 
     await repository.deleteProduct({
       id_product,
